refactor sudioctl auth impl slightly
