+++ src/cli/internal/cmd/auth/browser_login_internal_test.go
@@ -1,4 +1,4 @@
-package cmd
+package auth
 
 import (
 	"errors"
@@ -7,6 +7,8 @@ import (
 	"net/url"
 	"strings"
 	"testing"
+
+	"altinn.studio/studioctl/internal/config"
 )
 
 const (
@@ -65,12 +67,11 @@ func TestBuildStudioctlLoginURL(t *testing.T) {
 func TestResolveLoginTargetLocal(t *testing.T) {
 	t.Parallel()
 
-	command := NewAuthCommand(testConfig(t), nil)
-	target, err := command.resolveLoginTarget(loginFlags{env: "local"})
+	target, err := NewService(&config.Config{Home: t.TempDir()}).ResolveLoginTarget("local")
 	if err != nil {
-		t.Fatalf("resolveLoginTarget() error = %v", err)
+		t.Fatalf("ResolveLoginTarget() error = %v", err)
 	}
-	if target.scheme != testHTTP || target.host != testLocalHost {
+	if target.Scheme != testHTTP || target.Host != testLocalHost {
 		t.Fatalf("target = %+v, want http studio.localhost", target)
 	}
 }
@@ -78,10 +79,9 @@ func TestResolveLoginTargetLocal(t *testing.T) {
 func TestResolveLoginTargetUnknownEnvironment(t *testing.T) {
 	t.Parallel()
 
-	command := NewAuthCommand(testConfig(t), nil)
-	_, err := command.resolveLoginTarget(loginFlags{env: "unknown"})
+	_, err := NewService(&config.Config{Home: t.TempDir()}).ResolveLoginTarget("unknown")
 	if !errors.Is(err, ErrUnknownEnvironment) {
-		t.Fatalf("resolveLoginTarget() error = %v, want %v", err, ErrUnknownEnvironment)
+		t.Fatalf("ResolveLoginTarget() error = %v, want %v", err, ErrUnknownEnvironment)
 	}
 }
 
@@ -120,78 +120,10 @@ func TestLoginCallbackHandlerCancelled(t *testing.T) {
 	}
 
 	err := <-errCh
-	if !errors.Is(err, errLoginCancelled) {
-		t.Fatalf("error = %v, want %v", err, errLoginCancelled)
+	if !errors.Is(err, ErrLoginCancelled) {
+		t.Fatalf("error = %v, want %v", err, ErrLoginCancelled)
 	}
 	if len(codeCh) != 0 {
 		t.Fatal("unexpected login code")
-	}
-}
-
-func TestReadGitCredentialRequest(t *testing.T) {
-	t.Parallel()
-
-	request, err := readGitCredentialRequest(strings.NewReader(
-		"protocol=" + testHTTPS + "\nhost=" + testStudioHost + "\npath=repos/org/repo.git\n\n",
-	))
-	if err != nil {
-		t.Fatalf("read git credential request: %v", err)
-	}
-	if request.Protocol != testHTTPS || request.Host != testStudioHost || request.Path != "repos/org/repo.git" {
-		t.Fatalf("request = %+v", request)
-	}
-}
-
-func TestMatchesGitCredentialRequest(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name    string
-		request gitCredentialRequest
-		want    bool
-	}{
-		{
-			name:    "repos clone path",
-			request: gitCredentialRequest{Protocol: testHTTPS, Host: testStudioHost, Path: "repos/org/repo.git"},
-			want:    true,
-		},
-		{
-			name:    "leading slash repos clone path",
-			request: gitCredentialRequest{Protocol: testHTTPS, Host: testStudioHost, Path: "/repos/org/repo.git"},
-			want:    true,
-		},
-		{
-			name:    "wrong host",
-			request: gitCredentialRequest{Protocol: testHTTPS, Host: "example.com", Path: "repos/org/repo.git"},
-			want:    false,
-		},
-		{
-			name:    "wrong protocol",
-			request: gitCredentialRequest{Protocol: "http", Host: testStudioHost, Path: "repos/org/repo.git"},
-			want:    false,
-		},
-		{
-			name:    "non repos path",
-			request: gitCredentialRequest{Protocol: testHTTPS, Host: testStudioHost, Path: "org/repo.git"},
-			want:    false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-			if got := matchesGitCredentialRequest(tt.request, testHTTPS, testStudioHost); got != tt.want {
-				t.Fatalf("matchesGitCredentialRequest() = %t, want %t", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestMatchesGitCredentialRequestHTTP(t *testing.T) {
-	t.Parallel()
-
-	request := gitCredentialRequest{Protocol: testHTTP, Host: testLocalHost, Path: "repos/org/repo.git"}
-	if !matchesGitCredentialRequest(request, testHTTP, testLocalHost) {
-		t.Fatal("expected local http credential request to match")
 	}
 }
